L1P Phase 2 Bugs4 - Adjusted focus FWHM parameters to discard less stars
Based on forum feedback on using the new FWHM focus option, filtering of stars in the FWHM algorithm is too strict. The original parameters work fine on the sim but appear to be too restrictive for real world use. This change slackens off the filtering.

These parameter settings work on the Sim but need real world and may need further adjustment.

Whilst testing this, I found a bug in calculateR2 for Gaussian curves. Fix for this is included here.