Switch introduced between dynamic and static filter setting for alignment
When using the manual filter module the align module should always use the filter currently selected for capturing. On the other hand, when using an EFW it is in most cases better using a statically selected filter (typically the luminance filter).

To resolve this, a new switch is introduced in the Align module giving the user the option to select between using the currently selected filter or selecting statically a certain filter.

This addresses the issue #245 and is based on the first attempt of !868.