Speedup scheduler with large number of jobs
When there are many scheduler jobs (or a moderate number with a slow cpu), scheduler processing slows down. I tracked this down to an excessive/unnecessary number of calls to update graphics--both updating the same things over and over, as well as updating things that don't appear on the screen. This update fixes some of those issues.