Add DBus methods to query what KStars is currently centered on
There seems to be at present no way to determine the focus of KStars via DBus. There is a method called `getOption` which is curiously not exposed in the DBus interface in `org.kde.kstars.xml`, and is therefore not visible to clients. Moreover, a generic `getOption` method has the problem of not knowing which options are available. Therefore, this MR deprecates the `getOption` method (which does not seem to be used at all and is a candidate for removal) and instead introduces a `getFocusInformationXML` which returns comprehensive information about the focus as an XML object, consistent with other such DBus methods that return XML.