Refactoring Capture | extracting process steps (continued)
Capture still contains substantial parts of business logic making it very difficult to understand. The overall goal is to separate the business logic from the main (UI oriented) Capture class and to transform the code with hundreds of if-statements into state based logic.

This step is still part of the extraction phase, shifting code from Capture to the state machine CaptureModuleState and to the process container CaptureProcess:
- start() and stop()
- setting and removing devices (including their specific attributes)
- helper functions for various job attributes

With this change, processing steps directly implemented in the Capture module are transferred. Handling sequence jobs is the (hopefully) last part to be extracted, which will be subject to a separate change.