Image Overlay Component.
With this new feature, a user can add his/her own processed/completed astro-images, and the system will display them scaled and rotated appropriately on the SkyMap.

The feature is controlled in the KStars Settings menu, in a new tab labelled "Image Overlays". First the user needs to add files into a directory, parallel to the logs directory, called imageOverlays. Simply add the images there (typicaly jpegs, I suppose). Ideally these aren't massive files for performance reasons, but probably width 1000 or 2000 are fine. I have been testing with larger files, which will also work be use more system resources on slower CPUs.

The user then uses the Image Overlays menu in KStars Settings to (one-time) plate-solve the images and check a box to enable the image display. Successful plate-solve info is stored in the user-db so that it doesn't have to be done again. The images should, from then on, appear in the SkyMap in the proper position. There is a way to easily navigate to the images without manipulating the SkyMap by selecting a row in the overlay table and clicking on the "Show" button. You can move from one image to the next with up/down arrow keyboard commands.

A user can adjust the plate-solve timeouts. As these are mostly blind solves (jpegs won't have any header info, and as currently implemented, no header info is used) the plate solving can be problematic. You can choose a default image scale (arcseconds-per-pixel) or leave that to 0.0 to not use scale. If there are files that won't solve, the user can add RA,DEC into the image's row in the table displayed, which would get the solver to use the sky position as a constraint. The user can also add the scale that way. In fact, if the user knows all the info for the image, he/she can populate all the fields on the image's row and simply set the status field to OK, and plate-solving would no longer be required. 

To remove images from Image Overlays, you just need to remove them from the imageOverlays directory, and the next time KStars starts the image metadata will also be removed from the user-DB.

Testing before submit.

Here are instructions for those who'd like to test before submit.

1- Move a set of completed images to ~/.local/share/kstars/imageOverlays (create the directory if necessary).
   If you don't want to use your own, you can grab a set of 67 of my images from this drive directory:
   https://drive.google.com/drive/folders/10vg5vVQ6zNLl9TvdRA2JISHEhmj9LVgv?usp=sharing
   This directory only gets looked at when KStars starts, so when you add images you need to start or restart KStars.

2- Compile the KStars branch image "imageoverlay1" from https://invent.kde.org/murveit/kstars

3- Start KStars, go to Settings -> Configure KStars, and use the Image Overlays tab.

4- Check "Show Image Overlay" and choose a value for Max image dimension
   (1000 is ok to start with, but you can use a large value too).

5- Overlays are not visible until they are plate solved, which is a 1-time
   (pain in the butt) thing.
   You should be seeing the 67 images in the table, though none should say "Status: OK".
   Click on the 1st image, then shift-click on the last, then click "Solve".
   It should start solving all the images, with a timeout.
   This may take a while, primarily because of the ones that timeout after 30s.
   If you don't want to take too much time, just transfer 10 files into the directory.
   It's OK if some of the images didn't get solved (something to work on next
   is how to deal with that). For me 10-15 of the 67 don't solve.
   If you want to try and help an image solve, play with the RA/DEC/Scale in the image's row, and play 
   with the choice of solver profile.

6- You should now be able to browse the SkyMap and find the images.
   If you want to short-cut to them, just click on the image row
   in the table (assuming you've kept "center map on selection" checked)
   and it will move the skymap to the image's position and hopefully show the image.
   You should be able to zoom in/out, and pan around with reasonably quick response.