Refactoring Capture | extracting sequence job handling
This part of refactoring of the Capture module concentrates on decoupling sequence job handling from `Capture`:
- all business logic handling sequence jobs is moved from `Capture` to `CaptureProcess`
- making `SequenceJob` UI independent, updating the job table of the `Capture` UI moved to `Capture`

Additionally, handling of sequence jobs for previews and for calculating and displaying the image path is simplified. Creating these types of sequence jobs no longer adds and removes them from the job list, they are now kept separated.

This step (hopefully) finishes the separation of UI and business logic for the capture module:
- `Capture` manages the UI and is the DBus interface
- `CaptureModuleState` holds all attributes that define the state of capturing
- `CaptureProcess` holds the business logic
- `CaptureDeviceAdaptor` is the interface to the INDI devices making the state and business logic INDI independent
- `RefocusState` and `MeridianFlipState` are dedicated state machines for autofocus and the meridian flip
- `SequenceJob` holds all attributes relevant for a sequence of identical captures, its state is managed by `SequenceJobState`.