Bugfix  mount altitude limits
This MR fixes an issue with altitude limits in the Mount module.

The checkbox controlling the usage of limits is not honored, and it is not possible to choose negative altitudes, in case some target happens to be lower than the local horizon.

As the checkbox state has no counterpart in the INDI mount controller, the MR pushes the values marking invalid altitudes outside of [-90,+90] degrees and uses -91 and +91 degrees for low and high altitude respectively. Thus when altitude limits are disabled by the checkbox, any altitude is acceptable.