Take care of standard range for PA when creating a sequence job
All modules of KStars related to mosaic tiles (and some third party application like "Telescopius") use a full circle range for the position angle. So it is required to transform the angle to the de facto range of [-180° to 180°]in KStars while creating a sequence job.

see: [https://www.indilib.org/forum/ekos/13899-mosaics-and-position-angle.html](https://www.indilib.org/forum/ekos/13899-mosaics-and-position-angle.html)