Make "Set Coordinates Manually" dialog more intuitive
**Overview**

Makes the behavior of the "Set Coordinates Manually" dialog more intuitive by centering the display on the coordinates in the tab that is active when the user clicks "OK."

The existing behavior is to always use the RA/Dec coordinates if they can be parsed and only use the Az/Alt tab coordinates if RA/Dec parsing fails. This behavior is not intuitive, since clicking "OK" with the Az/Alt tab active can result in centering at a completely different location from the valued just entered on that tab.

**Testing**

I've tested this manually before and after the change using the following procedure:

1. Open Pointing \> Set Coordinates Manually 
2. Note or set the RA/Dec coordinates
3. Switch to Az/Alt tab
4. Set Az/Alt coordinates (choose something that is obviously a different location in the sky from the RA/Dec coordinates)
5. Click OK

Without this change, the above procedure would center on the coordinates in the RA/Dec tab. With this change, the sky is centered on the coordinates in the Az/Alt tab.

**Other Potential Improvements**

The program doesn't display any warning or error message if the coordinates can't be parsed. It just silently rejects them and does nothing. This isn't the worst, but would be better to give the user better feedback that the coordinates were rejected and maybe even why. This isn't addressed by this merge request.

**Bugzilla Ticket**

https://bugs.kde.org/show_bug.cgi?id=475499.