DSLR Focus Movement Bug
The bug is described in this thread:
https://indilib.org/forum/ccds-dslrs/13764-canon-dslr-focus-module.html

In essence, on a Canon DSLR the focus position as displayed on the Focus tab, does not keep count with the focuser movement. The focuser is a "relative" focuser and both in and out movements are adding to the current focus position rather than inward movements subtracting and outward movements adding. The reason for this appears to be that the focuser is returning an unsigned position and the code is using a state variable to work out whether the last movement was in or out. This state variable is not maintained correctly... hence the bug.

So the fix is to maintain the state variable correctly.