Improvements to the exposure calculator
Exposure Calculator Dialog Window is now resizeable

Fixed the display of tool tips (which were being masked by a tooltip that was placed on the frame)

Added an indicator for the sensor type of the selected camera. (Since camera names my not make that obvious)

Changed the limits of the "Noise Increase" input parameter to allow user to make larger over-ride adjustments of the computed sub-exposure.

Enabled support for cameras with static read noise. (Examples ATIK 16200 CCD and FLI 16200 CCD: these are CCD camera which do not have gain adjustments, so read noise is not variable).

Enabled direct download of camera data files from KStars Git hub to the user local kstars cameradata folder. (Note: users who have created their own camera data files, may wish set the file attribute to read-only, and/or make a back up copy) in case of an accidental over-write from the download function if the camera file has the same name).

Added an "experimental" graphical tool with the presentation of the Integration time / noise ratio curve, in an attempt to assist the user in determining an appropriate number of sub-exposures for integration. (A PDF explaining this tool will be posted to the dev forum, but changes to the user guide will be started soon).

Added several new camera data files:

```plaintext
    Atik-16200CCD_Mono.xml
    FLI-16200CCD_Mono.xml
    QHY_CCD_294M_Pro.xml
    QHY_CCD_461_PH.xml
    QHY_CCD_163C.xml
    QHY_CCD_163M.xml
    QHY_CCD_268C.xml
    QHY_CCD_294M.xml
    QHY_CCD_600_PH.xml
    ZWO_CCD_ASI294MC_Pro.xml
    ZWO_CCD_ASI294MM_Pro.xml
    ZWO_CCD_ASI533MC_Pro.xml
    ZWO_CCD_ASI2600MC_Pro.xml
    ZWO_CCD_ASI6200MC_Pro.xml
    ZWO_CCD_ASI533MC_Pro.xml
    ZWO_CCD_ASI533MM_Pro.xml
    Nikon_DSLR_DSC_D5100_(PTP_mode).xml   
    Nikon_DSLR_DSC_D700_(PTP_mode).xml

```