Refactor placeholder matching
Placeholders in the Capture Format string are serving for two purposes: (1) as matching string for existing files for counting frames and (2) as filename pattern to be filled when writing a new frame to a file.

On top of that, **pattern matching** is not identical for all parameters:
- For matching **%D** it is sufficient when the substring where %D is located forms a date/time expression
- If no target temperature is set, any number matches, but if the target temperature is set, only the concrete value counts
- Same is true for gain **%G**, offset **O**,  and ISO values **I** - if they are not explicitly set in the sequence job, any number matches, but if they are set, only the value from the sequence job creates a match
- The pier side **%P**, "East", "West" and "Unknown" match, no matter which pier side is currently active.
- All other values are mandatory currently, so only the value from the sequence job counts as match.

For generating a file name from a pattern, the current values should be used - even if the corresponding target value is not set for a sequence job (e.g. target temperature, target gain etc.).

With this fix, both the pattern matching is corrected handling sequence job parameters that might be optional (temperature, gain, offset) and the pattern filling for optional parameters is solved.