Revise differential slewing
A [contribution](https://indilib.org/forum/ekos/14052-rotator-doesn-t-re-attempt-pictures-automaticallly.html) in the forum mentions a problem with differential slewing.
The analysis of the program flow in align module shows a heavy flaw: Differential slewing ends prematurely, there is no iteration and the position threshold is never reached! (Just let the simulator do a differential slew and see how the position error remains red.)
This MR rectifies the program flow and corrects at the same time the above mentioned problem with the rotator.