Focus Guide Settle Bug
Bug in Focus, whereby the restart of guiding if the option to disable it was selected, was not followed by the specified pause before signalling Autofocus complete. The effect of this was to ignore the Guide Settle time after Autofocus before starting the next capture.

The bug was due to complicated code whereby a member variable was reset too early.