Minimal flip duration
Delays between starting a slew for a meridian flip and the mount reporting the slewing state might lead to misinterpretations of the mount's tracking state. This can happen since the mount state is executed regularly.

To avoid this, a tracking state is interpreted as the end of a slew earliest after configurable delay. This change introduces a new parameter for this delay and implements the delayed check.