Always show and edit scheduler job properties
With this change it is possible to display the properties of all scheduler jobs, even if the scheduler is running. Selecting an arbitrary scheduler job from the table displays its properties in the UI. This is possible after MR!1076 separated the job execution from the UI.

On top of this, it is possible to edit all jobs while the scheduler is running, even for the currently active one.
To enable this, there is one change necessary to the behavior of the Scheduler. Changes of job parameters become only effective when the save button has been pressed.

This is necessary to make resetting of job parameters possible by clicking on a different job row. The downside of this is, that the parameter changes aren't propagated to the job table. These values are updated as soon as the save button has been pressed, but not before.

For better readability, the currently running job is marked with a bold italic font:

![Bildschirmfoto_2023-12-28_um_20.15.59](/uploads/82ee02f3587c83a0101c1ec73774cdb3/Bildschirmfoto_2023-12-28_um_20.15.59.png)

The same logic has been applied to the capture module, where the currently running capture sequence is also marked bold and italic:

![Bildschirmfoto_2023-12-28_um_20.17.04](/uploads/7a6205a57070f02795278933f35468f6/Bildschirmfoto_2023-12-28_um_20.17.04.png)