Option making display of INDI window at startup configurable
When EKOS starts and connects to the INDI server, a window is created with a tab for each available INDI device. This change introduces a new option that makes it configurable, whether the INDI window should be shown when EKOS starts.

The option is enabled by default, i.e. the window opens as it does without this change.