Bug fix capture pausing blocks meridian flip
Pausing a capture sequence before the point has been reached that the meridian flip is already planned, works fine. But if the pause button is pressed in a capture sequence when the mount module is already waiting for the capture module to complete the current frame capture, then the meridian flip does not start.