Create standalone editor for editing and saving .esq files.
Currently you can only edit .esq files in the Capture tab, and when Capture/Scheduler is not running. With the new facility to edit running Scheduler jobs, it would be nice to also edit the .esq files. Its would also be nice to edit .esq files when not connected to INDI at all. This MR accomplishes both with a new stand-alone Sequence Editor. There are some compromises (there are aspects of .esq files that need information from connected cameras and filterwheels). This new editor uses the info from the most recent "live" Capture sessions. If that is not adequate then the user could wait until he/she can edit the .esq from the main Capture tab (when it is not busy).