Focus: Small tweaks as requested by forum user
Updated tooltip on Focus In / Out buttons to mention CCW (on In) and CW (on Out) which SCT users may be more familiar with.

When Focus In / Out or Goto are pressed, greyed out all three buttons until movement completes, when buttons are re-activated. This provides user feedback that the movement in progress until completion and prevents inadvertent multiple button presses.

Added max / min checks to Focus In / Out. This is more graceful than allowing moves to fail at the Indi level.