Add capture time delay estimate for rotator movements
The rotator utilities are extended with a routine that calculates a time delay estimate. This will be used as a time frame for which a capture is inhibited after a slew as long as the rotator is turning. In the optimal case EKOS now throws only one message about delaying capture.This is in contrast to the current situation where there can be dozens of messages of 10s delays depending on the rotator speed and the size of the rotation angle.

The idea emerged with [https://www.indilib.org/forum/ekos/13924-two-issues-with-meridian-flips.html](https://www.indilib.org/forum/ekos/13924-two-issues-with-meridian-flips.html)