Donut Buster - Phase 1
Improve focusing for telescopes with central obstructions that produce donut shaped stars when out of focus.

Background:
When out of focus, telescopes with central obstructions generate donut shaped stars. Star detection breaks down and fails to identify stars and when they are identified the HFRs are often unrealistic. This causes the Focus algorithm to fail.

Phase 1 is to make the Linear 1 Pass algorithm more tolerant of the poor data that it get when donuts are present. By identifying the good data in the "comfort zone" near focus, and excluding the wings where donuts are present and the data quality is poor, curvefitting can be better applied.

Usage Instructions:
1. Set Options->Process->Algorithm to Linear 1 Pass
2. Set Options->Process->Detection to SEP, Centroid or Gradient.
3. Set Options->Process->Measure to HFR, HFR Adj or FWHM.
4. Set Options->Process->Curve Fit to Hyperbola or Parabola
3. Set Options->Mechanics->Walk to Fixed Steps or CFZ Shuffle
4. Check Donut Buster:
![Donut_Buster](/uploads/8712ac9cea47ae95a0bfc537b8d684ad/Donut_Buster.png)

Donut Buster needs to start near focus so I suggest focusing firstly in the "normal way" to get the start point correct. Then check Donut Buster and rerun - probably not much will be different. Not increment the number of steps so the focuser moves further from optimum focus into donutland. If donut buster is working Autofocus should still draw and reasonable graph and find a good focus point.
![Donut_Buster_Example](/uploads/cc296ed7685fd57e80e9c5d39dff6abf/Donut_Buster_Example.png)