Separating Business Logic from UI in Scheduler | step 6
With this step, the main execution loop is moved to the process engine and the state machine:
- [x] Scheduler::stop()
- [x] Scheduler::start()
- [x] Scheduler::init()
- [x] Scheduler::iterate()
- [x] Scheduler::execute()
- [x] Scheduler::executeJob()
- [x] plus several functions used in the execution process

SchedulerProcess::removeJob() has a behavioral change, that after a removal, no line is selected. This is necessary to leave the deleted job's values in the UI, so that an immediate adding reverses the removal. This didn't work properly before.