Focus 2 Tick  Bugfix
The RainbowAstro focuser, unlike most focusers, moves to within 2 ticks of the requested position. Repeating a move request usually doesn't move the foccuser.

This causes a bug in Autofocus for walks = FIXED_STEP and CFZ_SHUFFLE.

This change mirrors the classic walk by allowing a tolerance of Initial Step Size in the datapoint position.