Separating Business Logic from UI in Scheduler | step 7
This part of refactoring concentrates on the DBus communication between the Scheduler and other modules. SchedulerProcess obtains its own DBus interface, which is required to communicate with signals through DBus with other modules.

The overall goal is shifting the entire DBus interface from Scheduler to SchedulerProcess, which additionally requires offering all methods from Scheduler, that could be accessed by DBus. This will be issue of an upcoming refactoring step.

Additionally, this change contains some minor changes:
- a minor bug in night time calculation has been fixed
- includes of scheduler.h in other modules has been optimized to improve compile time