New feature: Allow the user to mirror the sky map
With a new option in the "View" menu, the user will be able to mirror the sky map so as to be able to match the view through an erecting prism used for example on a Schmidt-Cassegrain or Refracting type telescope.

The rotation feature overlay now also marks East in addition to north and zenith, so as to know easily whether the display is mirrored or not.