Add edit profile button in FITS Viewer Solver and in Image Overlays
This adds an edit button next to the StellarSolver profile pull-down in the FITS Viewer Solver and in the ImageOverlay settings pages. These bring up the StellarSolver Profile Editor. In all cases it is editing the align profiles, which can also be edited in the Align tab.

There is also a bugfix to the file stellarsolverprofileeditor.cpp for an issue introduced in the recent donut-buster MR.