Focus Overscan Delay
This MR includes a new option "AF Overscan Delay" which introduces a user configurable delay between the outward movement and inward movement in an overscan move. There is also more logging of focuser movements.

Originally developed for a problem with the Sesto Senso focuser. This problem has now been fixed in the driver but this option could be useful in the future for other focusers that may have similar issues, such as the FLI_PDF focuser.