Expand check for initial rotator angle to IPS_OK & IPS_IDLE
In [this](https://www.indilib.org/forum/ekos/14150-sudden-problems-with-plate-solving-kstars-3-6-8.html?start=132#99684) thread about "differential slewing" a user of a "NiteCrawler" complains that the initial rotator angle at startup of the driver is not read in. The short check for the state of the corresponding field is therefore expanded from only "IPS_OK" to "IPS_OK & IPS_IDLE" to suit all drivers.