Location & time sources
This explicitly defines the master Time & Location sources that is used to sync both KStars + all connected INDI devices. The default time and location source is KStars. When an INDI Mount or GPS is connected, it is also available to act as the master source.

With this change, we remove any ambiguity on what controls the master location and time in Ekos.