Fix countdown
Small bugfix. The two capture countdown displays (one on the capture page and one on the manager page) count down both the progress of the current capture as well as counting down the total time left to complete the sequence job. However, when capturing previews there is no sequence and we should only count down the current capture.