Enable Qt6 in cmake configurations
As a first step towards Qt6 support of KStars, the cmake configurations are adopted. A new variable  "**BUILD_QT5**" is introduced to switch between the Qt5 and Qt6 based configurations. BUILD_QT5 is on by default so there should be no impact for the current Qt5 based builds.

For working on the Qt6 port (necessary in subsequent changes), set this variable to "Off" when running cmake.