Constellation boundaries actually precess along with the coordinate grid
The boundaries are defined by certain lines that are along circles of
fixed RA and of fixed Dec in the B1875 epoch. As the precession
correction is a rigid rotation of the equatorial coordinate grid,
these boundaries continue to remain along circles of fixed RA or
circles of fixed Dec, but the fixed RAs and Decs precess along with
the constellations.

I spot-checked that the boundaries rendered by KStars match standard
boundaries drawn on atlases, so it appears that the cbounds.dat we use
has J2000.0 entries. Thus, as long as we precess them from J2000.0 to
JNow, we will continue to render them correctly.

The checking of whether a point belongs in a constellation should
therefore use the present (RA, Dec) and not ICRS. We assume the
supplied point has been updated through updateCoords(). However,
proper motion allows stars to drift across constellation
boundaries. The correct order of corrections is:

Catalog Coordinates @ Epoch = J2000, Equinox = J2000
                    |
                    |
   Correct for Proper Motion from J2000 -> JNow
                    |
                    V
[Star Coordinates @ Epoch = JNow, Equinox = J2000]
                    |
                    |
   Correct for Precession, Nutation J2000 -> JNow
                    |
                    |
                    V
Equatorial Coordinates @ Epoch = JNow, Equinox = JNow
                    .
                    .
                    .

It is the coordinates that are marked in [] that must be used for
constellation membership. One can verify this by looking at the case
of rho Aql: https://en.wikipedia.org/wiki/Rho_Aquilae

To effect this, we try our best by making a hack: Try to convert the
SkyPoint into a StarObject. If it works, we apply the proper motion
and then test for constellation membership.