Bugfix for captured flat frames counting
Counting flat frames was not correct since the matching calibration frame wasn't counted in the captured frames map, but within the job itself. This stepped into another bug, where reducing the counter of captured frames did not take into account that the counter is an unsigned int.

Now both is fixed: the counters are updated after the decision what to do with the calibration frame, and the counter reduction is safe for unsigned integers.