Hanging bug due to multiple adjust focus requests
Bug reported by Magnus on the forum whereby, sporadically, processing will hang. From the log it looks like multiple adjust filter offset requests are coming in. My guess is that Align and Capture and both trying to adjust things that require a focus offset change - not enough evidence yet.

This MR will wait 3 x 10secs for the current in-flight focus request to complete before retrying. If there is still a problem the request will be aborted which should notify the caller of the issue, rather than just hanging.

Can't reproduce the bug reliably on the Sims but my best guess is that something is happening between Capture requesting a filter change (that involves offsets) and Align requesting a filter change. This results in Filter Manager requesting Focus to adjust the offset (whilst another filter offset request is still be processed). The second request was ignored resulting is the originator not receiving a "done" signal and therefore just hanging.

So the plan is to merge this and see if it fixes the problem.