Multiple cameras | step 6
With step we shift 100% of the capture process, state logic and device adapter to the camera level, since with the current single camera implementation, there is no big difference.

The rationale behind is that in a multi camera environment, each camera needs its own device adapter, state machine and process engine. The Capture module will then become an orchestrator of the capture processes running in parallel on several cameras. Therefore, in future steps of the multi camera support development, CaptureModuleState and CaptureProcess will be re-introduced.