Flag introduced restricting altitude limits to tracking
Some mounts (like the Avalons) have the tendency to pass altitudes below 0 even if the target is 0 or greater. This leads to the situation that altitude limits for such mounts are too restrictive.

For such situations, a new flag "Tracking only" is introduced for the altitude limits. If the flag is set (which is **not** the default), the altitude limits are only used if the mount is in a tracking state. While slewing or parking, the limits are ignored.